requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
pillow
jupyter
IPython
ipywidgets